Tweak the tagline
The current tagline needs a bit of a change as it suggests that actual 3D printers are welcome at the meetings. As 3D printer enthusiasts fall firmly under the scope of "maker", lets just remove that entirely.

Replace "all" with "other" as it makes it read a bit better to me.

Replace "folks" with "folk" as this can also be used as the plural and also reads a bit better to me.